+++ test/functional/android/driver_test.rb
@@ -50,7 +50,7 @@ class AppiumLibCoreTest
       error = assert_raises ::Selenium::WebDriver::Error::StaleElementReferenceError do
         e.click
       end
-      assert_equal 'does not exist', error.message
+      assert error.message.include?('does not exist')
       @driver.back
     end